Send the channel process instead of the channel id
If I want to be able to respond to a message I usually need the channel
process, not the id. If I really need the id, I can get it from the info
which I received during the connect. The other way round, one cannot get
the right channel based on the channel id alone, since the channel id
may be the same for multiple peers.

Closes #6